[FLINK-33098][autoscaler] Move non-kubernetes related autoscaler classes to flink-autoscaler module
