+++ flink-autoscaler/src/main/java/org/apache/flink/autoscaler/metrics/EvaluatedScalingMetric.java
@@ -15,29 +15,25 @@
  * limitations under the License.
  */
 
-package org.apache.flink.kubernetes.operator.autoscaler.metrics;
+package org.apache.flink.autoscaler.metrics;
 
-import org.apache.flink.runtime.rest.messages.job.metrics.AggregatedMetric;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
-import java.util.function.Function;
+/** Evaluated scaling metric. */
+@Data
+@NoArgsConstructor
+public class EvaluatedScalingMetric {
+    private double current;
 
-/** Enum specifying which aggregator to use when getting a metric value. */
-public enum MetricAggregator {
-    AVG(AggregatedMetric::getAvg),
-    MAX(AggregatedMetric::getMax),
-    MIN(AggregatedMetric::getMin);
+    private double average;
 
-    private final Function<AggregatedMetric, Double> getter;
-
-    MetricAggregator(Function<AggregatedMetric, Double> getter) {
-        this.getter = getter;
+    public EvaluatedScalingMetric(double current, double average) {
+        this.current = ScalingMetrics.roundMetric(current);
+        this.average = ScalingMetrics.roundMetric(average);
     }
 
-    public double get(AggregatedMetric metric) {
-        if (metric != null) {
-            return getter.apply(metric);
-        } else {
-            return Double.NaN;
-        }
+    public static EvaluatedScalingMetric of(double value) {
+        return new EvaluatedScalingMetric(value, Double.NaN);
     }
 }
